+++ gulp/handlebars.js
@@ -7,11 +7,9 @@ var livereload = require('gulp-livereload');
 var path = require('path');
 var streamqueue = require('streamqueue');
 var stripbom = require('gulp-stripbom');
-var htmlreplace = require('gulp-html-replace');
 
 var paths = require('./paths.js');
-
-var hiddenOptionsMeta = require('./hidden-option-meta');
+var metaOption = require('./meta-option');
 
 gulp.task('handlebars', function() {
 
@@ -20,9 +18,7 @@ gulp.task('handlebars', function() {
         '!*/**/*Partial.*'
     ])
         .pipe(stripbom({ showLog : false }))
-        .pipe(htmlreplace(hiddenOptionsMeta(), {
-            keepUnassigned: true
-        }))
+        .pipe(metaOption())
         .pipe(handlebars())
         .pipe(declare({
             namespace   : 'T',
