--- gulp/hidden-option-meta.yml
@@ -1,2 +1,0 @@
-general:
-  - bindAddress: false
